microservices/backend/requirements.txt
# List of Python dependencies
Flask==2.0.2
Flask-SocketIO==5.1.1
gevent==21.1.2
opencv-contrib-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
Werkzeug==2.0.2
gunicorn==20.1.0
python-socketio==5.4.1
gevent-websocket==0.10.1  
pymongo==4.0.1